address some MSVC warnings

+++ core/store/store_utils.cpp
@@ -275,7 +275,7 @@ uint32_t write_block(
 uint32_t write_block(
     data_output& out,
     const uint64_t* RESTRICT decoded,
-    uint32_t size,
+    uint64_t size,
     uint64_t* RESTRICT encoded) {
   assert(size);
   assert(encoded);
